Move soil standardise code out of soil models.
