+++ Models/Core/Simulation.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Data;
+using Models.Soils;
 
 namespace Models.Core
 {
@@ -210,7 +211,7 @@ namespace Models.Core
                 // Standardise the soil.
                 var soils = FindAllDescendants<Soils.Soil>();
                 foreach (Soils.Soil soil in soils)
-                    soil.Standardise();
+                    soil.Sanitise();
 
                 CheckNotMultipleSoilWaterModels(this);
 
@@ -272,8 +273,8 @@ namespace Models.Core
         /// <summary>
         /// Runs the simulation on the current thread and waits for the simulation
         /// to complete before returning to caller. Simulation is NOT cloned before
-        /// running. Use instance of Runner to get more options for running a 
-        /// simulation or groups of simulations. 
+        /// running. Use instance of Runner to get more options for running a
+        /// simulation or groups of simulations.
         /// </summary>
         /// <param name="cancelToken">Is cancellation pending?</param>
         public void Run(CancellationTokenSource cancelToken = null)
@@ -281,7 +282,7 @@ namespace Models.Core
             IsRunning = true;
             Exception simulationError = null;
 
-            // If the cancelToken is null then give it a default one. This can happen 
+            // If the cancelToken is null then give it a default one. This can happen
             // when called from the unit tests.
             if (cancelToken == null)
                 cancelToken = new CancellationTokenSource();
