+++ Tests/UnitTests/IrrigationTests.cs
@@ -22,7 +22,7 @@ namespace UnitTests
             var events = new Events(zone);
             var args = new object[] { this, new EventArgs() };
             events.Publish("Commencing", args);
-            events.Publish("StartOfSimulation", args); 
+            events.Publish("StartOfSimulation", args);
             events.Publish("DoDailyInitialisation", args);
 
             var soilWater = zone.Children[1].Children[7] as Models.WaterModel.WaterBalance;
@@ -121,7 +121,7 @@ namespace UnitTests
                                 Thickness = new double[] { 100, 300, 300, 300, 300, 300 },
                                 InitialValues = new double[] { 0, 0, 0, 0, 0, 0 },
                                 InitialValuesUnits = Solute.UnitsEnum.kgha
-                            },                             
+                            },
                             new Water()
                             {
                                 Thickness = new double[] { 100, 300, 300, 300, 300, 300  },
@@ -193,7 +193,7 @@ namespace UnitTests
             events.ConnectEvents();
 
             var soil = zone.Children[1] as Soil;
-            soil.Standardise();
+            soil.Sanitise();
 
             return zone;
         }
